Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 int main() {
     
-    float num1;
+    int num1;
     ;
     scanf("%f",&num1);
-    float num2;
+    int num2;
     
     scanf("%f",&num2);
-    float num3;
+    int num3;
     
     scanf("%f",&num3);
